Update number of items.py
Added "\n" between contestant names to make it easier to see all names. Fixed the coding for obtaining the amount of names. Before it was giving me a weird error that my indent was wrong no matter what I did so I made it into a function and added more parameters. Also I added more ways for incorrect entries to get feedback, most of which allow the user to start over automatically.